Allow developers to specify the base URL to the wp-forms-api library
Unfortunately, plugins_url() is not a reliable function to use in our case:

* The returned URL can be modified using a filter which may not be where the
library actually exists.
* If the library is included in a theme (and not a plugin) it may return an
incorrect path.
* If symlinks are used, it may return an incorrect path.

This commit adds the "set_url" method to the WP_Forms_API class which
allows developers to, optionally, override the the base URL to wp-forms-api
directory if they are experiencing issues with CSS and JS files not loading
due to invalid paths.

@see https://make.wordpress.org/core/2014/04/14/symlinked-plugins-in-wordpress-3-9/
@see @see https://core.trac.wordpress.org/ticket/13550