Update checkbox handling logic
The previous way this was handled, checkboxes were forced to have the value of `true`, meaning the checkbox would only function properly if the `'#checked'` property was set to boolean `true`. This allows an arbitrary value to be used as the value to store for the checkbox, and an arbitrary value to be used in the value attribute of the element (passed through `'#attrs' => [ 'value' => 'foo' ]`).